refactor: 이벤트 핸들러 파일명 수정 및 이벤트 관련 파일 관심사 통합
- typeGuards 파일에 있던 isElements는 이벤트 타겟에 대한 타입 가드이므로 events 폴더 내부로 옮김
- api 폴더 구조와 통일되도록 함
