+++ src/events/guards.ts
