+++ src/main.ts
@@ -1,8 +1,9 @@
 import { Button } from "./components/index";
 import { renderMoviesList } from "./domains/renderMoviesList";
-import "./events/index";
+import "./events/handlers";
 
 addEventListener("load", async () => {
+  // TODO: 더 보기 버튼 삭제 예정 (2단계에서 무한 스크롤로 대체)
   const $container = document.querySelector(".container");
 
   $container?.appendChild(
